SEPP-34 Refactor MongoModel
