+++ dating_app/schemas/ChatModel.py
@@ -1,32 +1,7 @@
-from bson import ObjectId
 from fastapi import Body
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
-
-# https://www.mongodb.com/developer/quickstart/python-quickstart-fastapi/#the-_id-attribute-and-objectids
-class PyObjectId(ObjectId):
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, v):
-        if not ObjectId.is_valid(v):
-            raise ValueError("Invalid objectid")
-        return ObjectId(v)
-
-    @classmethod
-    def __modify_schema__(cls, field_schema):
-        field_schema.update(type="string")
-
-
-class MongoModel(BaseModel):
-    id: PyObjectId = Field(default_factory=PyObjectId, alias="_id")
-
-    class Config:
-        allow_population_by_field_name = True
-        arbitrary_types_allowed = True
-        json_encoders = {ObjectId: str}
+from dating_app.schemas.MongoModel import MongoModel
 
 
 class InitiateChat(MongoModel):
